+++ src/main/java/br/com/hellobank/api/service/interfaceServ/ISaqueService.java
@@ -2,12 +2,12 @@ package br.com.hellobank.api.service.interfaceServ;
 
 import java.util.List;
 
-import br.com.hellobank.api.model.entity.Saque;
+import br.com.hellobank.api.model.entidades.Saque;
 import br.com.hellobank.api.model.request.SaqueRequest;
 
 public interface ISaqueService {
 
     Saque create(SaqueRequest saqueRequest);
 
-    List<Saque> findClienteSaques(Long id);
+    List<Saque> findClienteSaques(Integer id);
 }
