+++ src/main/java/br/com/hellobank/api/service/interfaceServ/ITransferenciaService.java
@@ -2,12 +2,12 @@ package br.com.hellobank.api.service.interfaceServ;
 
 import java.util.List;
 
-import br.com.hellobank.api.model.entity.Transferencia;
+import br.com.hellobank.api.model.entidades.Transferencia;
 import br.com.hellobank.api.model.request.TransferenciaRequest;
 
 public interface ITransferenciaService {
 
 	Transferencia create(TransferenciaRequest transferenciaRequest);
 
-    List<Transferencia> findClienteTransferencia(Long id);
+    List<Transferencia> findClienteTransferencia(Integer id);
 }
